feat: Refactor code in 04-homework.ts file
The code in the 04-homework.ts file has been refactored to simplify the functions and improve readability. The isRedFruit function now uses the includes method of an array to check if the fruit is red. The getFruitsByColor function has been updated to use a switch statement or an object literal to return the fruits based on the color. The workingSteps function has been simplified using conditional statements.